Create index.js
This the node.js that is used as the back-end for the alexa skill. It uses the characteristic from Fermat's Little Theorem that states that if the modulo is prime(p), the the congruence class is p-1. Using a linear search, the best prime is picked based on the size of the adjective list.